Combine form_args from extension and user space
This lets the user space add things like validators to a field, while
still benefiting from the widget provided by the extension.

This provides merge logic for the top-level keys only, overriding any
keys that are present from the extension. If a key is present in both
the extension and the user space, the user space will override
it entirely.